Move the `get_id_token_payload()` function to apps.users.utils (more appropriate location)
